+++ src/main.ts
@@ -45,13 +45,6 @@ let CYOL = {
 
     // Used by Game.registerMod
     init: function() {
-        // Legacy data, was previously stored in CCSE.config.OtherMods
-        if(CCSE.config.OtherMods.CYOL) {
-            loadSettingsFrom(JSON.stringify(CCSE.config.OtherMods.CYOL));
-            // Using JSON.stringify is easier than writing a separate function just for legacy support
-            delete CCSE.config.OtherMods.CYOL; // be a good citizen and not bloat CCSE's save object
-        }
-
         Game.customLumpTooltip.push(customLumpTooltip);
         Game.customOptionsMenu.push(customOptionsMenu);
         Game.customStatsMenu.push(function() {